librarify zknotes server; embed into tauri app
