+++ server-lib/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "zknotes-server-lib"
+version = "0.1.0"
+authors = ["Ben Burdette <bburdette@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+zkprotocol = { path = "../zkprotocol" }
+orgauth = { path = "../orgauth/rustlib/" }
+env_logger = "0.5.13"
+rusqlite = "0.29.0"
+log = "0.4.0"
+actix-web = "4.3.1"
+actix-files = "0.6.2"
+actix-session = { version = "0.7.2", features = ["cookie-session"] }
+actix-cors = "0.6.4"
+actix-multipart = "0.6.0"
+futures = "0.3.28"
+futures-util = { version = "0.3.17", default-features = false, features = ["std"] }
+serde = "1.0.124"
+serde_derive = "1.0.124"
+serde_json = "1.0.64"
+uuid = { version = "0.8", features = ["serde", "v4"] }
+crypto-hash = "0.3.4"
+time = "0.2.27"
+rand = "0.5.0"
+toml = "0.5.9"
+simple-error = "0.2.1"
+json = "*"
+barrel = { version = "0.7.0", features = ["sqlite3"] }
+clap = "2.33.2"
+timer = "0.2.0"
+chrono = "0.4.26"
+either = "1.6.1"
+sha256 = "1.1.1"
+tracing-actix-web = "0.7"
+tracing = "0.1"
