+++ server-lib/src/interfaces.rs
