+++ server-lib/src/search.rs
